ref(utils): move issue-tag query helpers out of the utils barrel
`escapeIssueTagKey` and `generateQueryWithTag` were the only members of the
`sentry/utils` god-barrel that imported `types/event`, `utils/fields`, and
`utils/queryString` — the edges keeping the barrel inside the frontend
type-import strongly-connected component (SCC).

Relocate both into `utils/queryString` (alongside the related
`appendTagCondition`) and repoint the handful of consumers. The barrel then
imports nothing in the cycle and leaves the SCC. Re-exporting from the barrel
would re-create the edge (the helpers' new home is itself in the cycle), so
consumers import from `sentry/utils/queryString` directly.

Largest SCC drops 1867 -> 1850 standalone; the barrel is one of the most
widely-imported modules, so this compounds with the other de-cycling PRs.
